refactor: Router, App 분리
