+++ src/js/Components/NavBar/index.js
@@ -4,37 +4,14 @@ import Component from '../../core/Component';
 import template from './template';
 
 class NavBar extends Component {
-  constructor({ parentNode, state, props: { goPage, setIsLogin } }) {
+  constructor({ parentNode, state }) {
     super({ parentNode, state });
-
-    this.goPage = goPage;
-    this.setIsLogin = setIsLogin;
   }
 
   renderSelf() {
     this.parentNode.innerHTML = template(
       this.state.isLogin ? AUTHENTICATED_LINK : UNAUTHENTICATED_LINK
     );
-  }
-
-  addEventListeners() {
-    this.parentNode.addEventListener('click', (e) => {
-      const anchor = e.target.closest('.js-link');
-      if (!anchor) return;
-
-      e.preventDefault();
-
-      const path = anchor.getAttribute('href');
-      if (path === AUTHENTICATED_LINK.LOGOUT.PATH) {
-        localStorage.removeItem(LOCAL_STORAGE_KEY.ACCESSTOKEN);
-        this.setIsLogin(false);
-        this.goPage(UNAUTHENTICATED_LINK.LOGIN.PATH);
-
-        return;
-      }
-
-      this.goPage(path);
-    });
   }
 }
