+++ src/js/Pages/Station/EditModal/template.js
@@ -1,5 +1,5 @@
-export const stationModal = ({ modalKey }) => `
-  <div class="modal ${modalKey}-modal">
+export const stationModal = ({ modalName }) => `
+  <div class="modal ${modalName}-modal">
     <div class="modal-inner p-8">
       <button class="modal-close">
         <svg viewbox="0 0 40 40">
@@ -9,14 +9,14 @@ export const stationModal = ({ modalKey }) => `
       <header>
         <h2 class="text-center">🛤️ 역 수정</h2>
       </header>
-      <form id="${modalKey}-form">
+      <form id="${modalName}-form">
         <div class="input-control">
-          <label for="${modalKey}-name" class="input-label" hidden
+          <label for="${modalName}-name" class="input-label" hidden
             >역 이름</label
           >
           <input
             type="text"
-            id="${modalKey}-name"
+            id="${modalName}-name"
             name="name"
             class="input-field"
             placeholder="역 이름"
