--- src/js/core/PageComponent.js
@@ -1,7 +1,0 @@
-import Component from './Component';
-
-class PageComponent extends Component {
-  constructor() {
-    super(arguments);
-  }
-}
